requirements.txt
requests
python-dotenv
graypy